View sorgularının classları oluşturuldu ve bu viewlere ilişkin sorguların yapılabilmesi için ekstra sorgu classları açıldı ve ReportServices kalsörü güncellendi gerekli metotlar yazıldı.

+++ CalorieTrackingApp/CalorieTrackingApp.DAL/ValueObject/UserMealDailyFoodRecord.cs
@@ -12,7 +12,7 @@ namespace CalorieTrackingApp.DAL.ValueObject
     {
         public int UserId { get; set; }
         public string MealName { get; set; }
-        public float FoodCalorie { get; set; }
+        public int FoodCalorie { get; set; }
         public decimal Value {  get; set; }
         public DateOnly RecordDate { get; set; }
     }
